reverted expiration changes
