+++ pkg/controllers/deprovisioning/metrics.go
@@ -18,7 +18,7 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	crmetrics "sigs.k8s.io/controller-runtime/pkg/metrics"
 
-	"github.com/aws/karpenter/pkg/metrics"
+	"github.com/aws/karpenter-core/pkg/metrics"
 )
 
 func init() {
